refactor: lukemorales query key factory 적용
