+++ package.json
@@ -12,6 +12,7 @@
     "coverage": "vitest run --coverage"
   },
   "dependencies": {
+    "@lukemorales/query-key-factory": "^1.3.4",
     "@tanstack/react-query": "^5.64.2",
     "clsx": "^2.1.1",
     "react": "^19.0.0",
