+++ src/features/comment/lib/index.ts
@@ -1,0 +1,5 @@
+import { createQueryKeys } from "@lukemorales/query-key-factory"
+
+export const commentKeys = createQueryKeys("comment", {
+  fetch: () => ["comment"],
+})
