+++ src/features/comment/model/useCommentsQuery.ts
@@ -1,9 +1,10 @@
 import { useQuery } from "@tanstack/react-query"
 import { fetchCommentsByPostId } from "@entities/post/api"
+import { commentKeys } from "@features/comment/lib"
 
 export const useCommentsQuery = (postId: number) => {
   return useQuery({
-    queryKey: ["/api/comments/post"],
+    queryKey: commentKeys.fetch().queryKey,
     queryFn: () => fetchCommentsByPostId(postId),
   })
 }
