+++ src/features/post/model/useTagsQuery.ts
@@ -1,9 +1,10 @@
 import { useQuery } from "@tanstack/react-query"
 import { fetchTags } from "@entities/post/api"
+import { postKeys } from "@entities/post/lib"
 
 export const useTagsQuery = () => {
   return useQuery({
-    queryKey: ["/api/posts/tags"],
+    queryKey: postKeys.tags().queryKey,
     queryFn: () => fetchTags(),
   })
 }
